collettore/requirements.txt
# DarkFleet Collettore Dependencies
# ==================================

# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
starlette==0.41.3
python-dotenv==1.0.1
slowapi==0.1.9

# WebSocket
websockets==13.1

# Data validation
pydantic==2.9.2
pyyaml==6.0.2

# Redis
redis==5.1.1

# Logging
structlog==24.4.0
colorama==0.4.6

# Keycloak/JWT Authentication
python-keycloak==4.0.0
PyJWT==2.8.0
cryptography==44.0.0
itsdangerous==2.1.2
httpx==0.28.1

# Testing (optional)
pytest==7.4.4
pytest-asyncio==0.23.3
